+++ controller/app/system/system-safe-i2c.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
 void system_safe_i2c_init();
-bool system_safe_i2c_write(uint8_t address, uint8_t reg, uint8_t* buf, uint8_t buf_size);
-bool system_safe_i2c_read(uint8_t address, uint8_t reg, uint8_t* buf, uint8_t buf_size);
+bool system_safe_i2c_write(uint8_t address, uint8_t reg, uint8_t* buf, uint16_t buf_size);
+bool system_safe_i2c_read(uint8_t address, uint8_t reg, uint8_t* buf, uint16_t buf_size);
